backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
isodate==0.7.2
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
mcp==1.25.0
more-itertools==10.8.0
openai==2.15.0
platformdirs==4.5.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
simple-salesforce==1.12.9
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
zeep==4.3.2

